refactor : 그룹 등록, 해제 이벤트를 공통모듈에서 제거하고 ordertable 도메인으로 이동
