+++ kitchenpos-ordertable/src/main/java/kitchenpos/ordertable/domain/GroupingEventListener.java
@@ -1,8 +1,8 @@
 package kitchenpos.ordertable.domain;
 
-import kitchenpos.common.event.GroupEvent;
-import kitchenpos.common.event.GroupInfo;
-import kitchenpos.common.event.UngroupEvent;
+import kitchenpos.ordertable.event.GroupEvent;
+import kitchenpos.ordertable.event.GroupInfo;
+import kitchenpos.ordertable.event.UngroupEvent;
 import kitchenpos.ordertable.exception.DuplicateTablesException;
 import kitchenpos.ordertable.exception.IllegalGroupingTableStateException;
 import kitchenpos.ordertable.exception.NotEnoughTablesException;
