+++ kitchenpos-ordertable/src/main/java/kitchenpos/ordertable/event/GroupInfo.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.event;
+package kitchenpos.ordertable.event;
 
 import java.util.List;
