+++ kitchenpos-ordertable/src/main/java/kitchenpos/ordertable/event/UngroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.event;
+package kitchenpos.ordertable.event;
 
 import org.springframework.context.ApplicationEvent;
