+++ kitchenpos-tablegroup/src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -5,9 +5,9 @@ import kitchenpos.tablegroup.domain.TableGroupRepository;
 import kitchenpos.ordertable.dto.OrderTableRequest;
 import kitchenpos.tablegroup.dto.TableGroupRequest;
 import kitchenpos.tablegroup.dto.TableGroupResponse;
-import kitchenpos.common.event.GroupEvent;
-import kitchenpos.common.event.GroupInfo;
-import kitchenpos.common.event.UngroupEvent;
+import kitchenpos.ordertable.event.GroupEvent;
+import kitchenpos.ordertable.event.GroupInfo;
+import kitchenpos.ordertable.event.UngroupEvent;
 import kitchenpos.tablegroup.exception.TableGroupNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
